feat: relax tx commitment check
currently the node only sends an account hash update if the update happens in the block range it's sending data from. This poses an issue when syncing because if we send multiple TXs before syncing we might end up with some in an uncommitted state while it was actually committed. Since it's very unlikely that another transaction can generate the same output notes with the same account hash changes, instead of checking both we'll just accept one of those two matching our local TX.